server/requirements.txt
asgiref==3.11.1
dj-database-url==3.1.0
Django==6.0.2
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Markdown==3.10.1
psycopg==3.3.2
psycopg-binary==3.3.2
PyJWT==2.11.0
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0